pyproject.toml
[project]
name = "vlm-grpo-rw"
version = "0.1.0"
description = "GRPO training pipeline for reducing RW flips in VLM self-reflection"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2.0",
    "transformers>=4.50.0",
    "trl>=0.16.0",
    "peft>=0.14.0",
    "datasets>=3.0.0",
    "accelerate>=1.2.0",
    "Pillow>=10.0.0",
    "tqdm>=4.60.0",
]

[project.optional-dependencies]
vllm = ["vllm>=0.7.0"]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.0.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vlm_grpo"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.isort]
known-first-party = ["vlm_grpo"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
